+++ src/types/public-api-class-group.ts
@@ -1,9 +1,0 @@
-import { PublicApiClass } from './public-api-class.js';
-
-export interface PublicApiClassGroup {
-  name: string;
-  description?: string;
-  groups?: PublicApiClassGroup[];
-  classes?: PublicApiClass[];
-  imageToken?: string;
-}