+++ src/types/public-api-styles.ts
@@ -1,0 +1,7 @@
+import { PublicApiStyle } from './public-api-style.js';
+import { PublicApiStyleGroup } from './public-api-style-group.js';
+
+export interface PublicApiStyles {
+  groups?: PublicApiStyleGroup[];
+  styles?: PublicApiStyle[];
+}