+++ src/types/public-api-token.ts
@@ -1,6 +1,11 @@
+import { DemoMetadata } from './demo-metadata.js';
+
 export interface PublicApiToken {
   name: string;
-  customProperty: string;
+  customProperty?: string;
   description?: string;
-  deprecatedCustomProperty?: string;
+  deprecatedCustomProperties?: string[];
+  obsoleteCustomProperties?: string[];
+  cssProperty?: string;
+  demoMetadata?: DemoMetadata;
 }