+++ src/types/public-token-set.ts
@@ -1,4 +1,5 @@
 export type PublicTokenSet = {
   name: string;
   path: string;
+  deprecatedTokensPath?: string;
 };