+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0 (2026-03-17)
+- Expand support for public API custom properties and classes. [#6](https://github.com/blackbaud/skyux-branding-builder/pull/6)
+
+## BREAKING CHANGES
+- Rename PublicApiClass classes to PublicApiStyle, convert deprecatedCustomProperty and deprecatedClassName to lists
+
 # 1.1.0 (2026-03-04)
 - Add support for public API custom properties and classes. [#5](https://github.com/blackbaud/skyux-branding-builder/pull/5)